@@ -1,71 +1,62 @@
 {
   "name": "recoverysigner-firebase-auth",
-  "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@babel/core": "^7.10.5",
-    "@babel/helper-call-delegate": "^7.11.4",
-    "@babel/helper-validator-option": "^7.14.5",
-    "@lingui/babel-preset-react": "^2.9.1",
-    "@lingui/cli": "^2.9.1",
-    "@lingui/macro": "^2.9.1",
-    "@lingui/react": "^2.9.1",
-    "@sentry/browser": "^6.16.1",
-    "@stellar/eslint-config": "^1.0.3",
-    "@stellar/prettier-config": "^1.0.1",
-    "@stellar/tsconfig": "^1.0.2",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.0.4",
-    "@testing-library/user-event": "^12.1.6",
-    "@types/jest": "^26.0.4",
-    "@types/lingui__macro": "^2.7.3",
-    "@types/lingui__react": "^2.8.2",
-    "@types/lodash": "^4.14.150",
-    "@types/node": "^14.11.2",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-redux": "^7.1.7",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-plugin-macros": "^2.8.0",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-config-react": "^1.1.7",
-    "firebase": "^9.6.2",
-    "lodash": "^4.17.21",
-    "prettier": "^2.0.5",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-scripts": "3.4.3",
-    "reduce-reducers": "^1.0.4",
-    "typescript": "4.0.3"
-  },
-  "devDependencies": {
-    "firebase-tools": "^10.0.1",
-    "husky": "7.0.4"
+  "version": "1.0.0",
+  "type": "module",
+  "engines": {
+    "node": ">=22"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "yarn lingui:compile && react-scripts build",
+    "start": "yarn lingui:compile && vite",
+    "build": "yarn lingui:compile && tsc -b && vite build && rm -r build/config/",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "lingui:compile": "yarn lingui extract && yarn lingui compile",
+    "lingui:compile": "yarn lingui extract && yarn lingui compile --typescript",
     "lint": "eslint 'src/**/*.{js,jsx,ts,tsx}'",
     "lint:fix": "eslint 'src/**/*.{js,jsx,ts,tsx}' --fix",
     "prettier": "prettier --check 'src/**/*.{js,jsx,ts,tsx}'",
     "prettier:fix": "prettier --write 'src/**/*.{js,jsx,ts,tsx}'",
-    "prepare": "husky install"
+    "prepare": "husky",
+    "postinstall": "npx patch-package"
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+  "dependencies": {
+    "@lingui/core": "^5.4.1",
+    "@lingui/macro": "^5.4.1",
+    "@lingui/react": "^5.4.1",
+    "@sentry/browser": "^10.5.0",
+    "firebase": "^12.1.0",
+    "lodash": "^4.17.21",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-redux": "^9.2.0",
+    "reduce-reducers": "^1.0.4",
+    "redux": "^5.0.1"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.33.0",
+    "@lingui/cli": "^5.4.1",
+    "@lingui/vite-plugin": "^5.4.1",
+    "@stellar/prettier-config": "^1.2.0",
+    "@types/lodash": "^4.17.20",
+    "@types/react": "^19.1.10",
+    "@types/react-dom": "^19.1.7",
+    "@typescript-eslint/eslint-plugin": "^8.39.1",
+    "@typescript-eslint/parser": "^8.39.1",
+    "@vitejs/plugin-react": "^5.0.0",
+    "babel-plugin-macros": "^3.1.0",
+    "eslint": "^9.33.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-config-react": "^1.1.7",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "firebase-tools": "^14.12.0",
+    "globals": "^16.3.0",
+    "husky": "^9.1.7",
+    "import": "^0.0.6",
+    "prettier": "^3.6.2",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.39.1",
+    "vite": "^7.1.2"
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
