@@ -1,5 +1,5 @@
-import { Status } from "types.d/Status";
-import { State } from "types.d/State";
+import { type Status } from "types/Status";
+import { type State } from "types/State";
 
 type Actions = SetStatusAction;
 
